Split nested types into dedicated files
