+++ OfficeIMO.Word.Pdf/PdfFootnote.cs
@@ -1,0 +1,7 @@
+namespace OfficeIMO.Word.Pdf {
+    internal sealed class PdfFootnote {
+        public int Number { get; set; }
+        public string Text { get; set; } = string.Empty;
+    }
+}
+
